requirements.txt
feedparser>=6.0
Jinja2>=3.1
anthropic>=0.39
PyYAML>=6.0
python-dotenv>=1.0
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.2
beautifulsoup4>=4.12
